chore: move component from feature folder to pages folder
