+++ src/config/route.ts
@@ -3,7 +3,6 @@ const routePath = {
   login: "/login",
   bookingService: "/booking-service/:serviceId",
   bookingDetail: "/booking-detail/:orderId",
-  manage: "/manage",
   signup: "/sign-up",
   store: "/store",
   storeDetail: "/store-detail/:id",
