--- src/features/auth/hooks/useEmail.ts
@@ -1,3 +1,0 @@
-export const useEmail = () => {
-  //Custom hook for auth, use email name is temporary
-};
