--- src/features/home/hooks/useMind.ts
@@ -1,3 +1,0 @@
-export const useMind = () => {
-  //init custom hook for home
-};
