--- src/features/home/types/init.ts
@@ -1,4 +1,0 @@
-//Init type for home url
-export type Init = {
-  id: string;
-};
