+++ src/pages/home/_components/hero-section.tsx
