+++ src/pages/home/constants/hero-image.ts
