+++ src/pages/home/index.tsx
@@ -1,8 +1,8 @@
-import Hero from "@/features/home/components/hero-section";
-import Services from "@/features/home/components/services";
-import StoreCarousel from "@/features/home/components/store-carousel";
-import Feedback from "@/features/home/components/feedback";
-import CallToAction from "@/features/home/components/call-to-action";
+import Hero from "./_components/hero-section";
+import Services from "./_components/services";
+import StoreCarousel from "./_components/store-carousel";
+import Feedback from "./_components/feedback";
+import CallToAction from "./_components/call-to-action";
 
 const Home = () => {
   return (
