--- src/pages/manage/index.tsx
@@ -1,5 +1,0 @@
-const Manage = () => {
-  return <div>Manage</div>;
-};
-
-export default Manage;
