+++ src/pages/store/_components/options-dropdown.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 import { Button } from "@/components/ui/button";
 import { FILTER_OPTIONS, FILTER_CATEGORIES } from "../constants/filter";
-import type { FilterOption } from "../types/store.type";
+import type { FilterOption } from "@/features/store/types/store.type";
 import Icon from "@/components/icons";
 import SelectField from "@/components/common/select-field";
