--- src/providers/theme-provider.tsx
@@ -1,5 +1,0 @@
-const ThemeProvider = () => {
-  return <div>ThemeProvider</div>;
-};
-
-export default ThemeProvider;
