+++ src/routes/index.tsx
@@ -1,20 +1,18 @@
-import { Route } from "../types/routes.type";
+import type { Route } from "../types/routes.type";
+import { AuthRoute } from "./auth-route";
 import routePath from "../config/route";
 import { PrivateRoute } from "./private-route";
 import MainLayout from "../layouts/main-layout";
-import AdminLayout from "../layouts/admin-layout";
 import AuthLayout from "../layouts/auth-layout";
 import CreateOrder from "../pages/order/creation";
 import Login from "../pages/auth/login";
 import Home from "../pages/home";
-import Manage from "../pages/manage";
 import SignUp from "../pages/auth/signup";
-import StorePage from "../pages/store/page";
+import StorePage from "../pages/store";
 import StoreDetail from "../pages/store-detail";
 import ServiceDetail from "../pages/service/service-detail";
-import ManageStore from "../pages/manage-store";
+import ManageStore from "../pages/store/manage";
 import NotFound from "../pages/not-found";
-import { AuthRoute } from "./auth-route";
 import OrderTracking from "../pages/order/tracking";
 import Profile from "../pages/profile";
 
@@ -59,15 +57,6 @@ const listRoute: Route[] = [
       </PrivateRoute>
     ),
     layout: MainLayout,
-  },
-  {
-    path: routePath.manage,
-    component: (
-      <PrivateRoute>
-        <Manage />
-      </PrivateRoute>
-    ),
-    layout: AdminLayout,
   },
   {
     path: routePath.store,
