--- src/services/endpoint.ts
@@ -1,8 +1,0 @@
-export const ENDPOINTS = {
-  AUTH: {
-    LOGIN: "/auth/login",
-    REGISTER: "/auth/register",
-    LOGOUT: "/auth/logout",
-    REFRESH_TOKEN: "/auth/refresh-token",
-  },
-};
