--- src/utils/validation/authValidation.ts
@@ -1,5 +1,0 @@
-export const authValidation = {
-  validateEmail: (email: string) => {
-    return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
-  },
-};
